% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_FIA_db.R
\name{download_and_unzip}
\alias{download_and_unzip}
\title{download and unzip file from web}
\usage{
download_and_unzip(url, file_dir)
}
\arguments{
\item{url}{web address of .zip file to be downloaded}

\item{file_dir}{file path of directory into which file will be unzipped}
}
\value{
list of file paths extracted from .zip
}
\description{
download a file from a web address (`url`) and unzip it into a
specified directory (`file_dir`).
}
\examples{
\dontrun{
unzipped <- download_and_unzip(
  url = "https://apps.fs.usda.gov/fia/datamart/CSV/MN_PLOT.zip",
  file_dir = tempdir()
)
}
}
