% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_FIADB.R
\name{download_by_state}
\alias{download_by_state}
\title{download FIA database files by state}
\usage{
download_by_state(
  state,
  file_dir = tempdir(),
  files = c("PLOT", "SUBPLOT", "COND", "TREE", "SURVEY")
)
}
\arguments{
\item{state}{character state abbreviation code}

\item{file_dir}{directory into which files will be downloaded. By default
this is the R temporary directory (`tempdir()`)}

\item{files}{list of FIA tables to download. Tables must be identified by
their 'Oracle Table Name' as described in the 'Index of Tables' in FIADB
User Guide.}
}
\value{
list of filepaths
}
\description{
This function downloads FIA data for a single state
}
\examples{
\dontrun{
fia_files <- download_fia_by_state("MN")
}

}
\author{
Henry Rodman
}
