% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{plot.tidyFIA}
\alias{plot.tidyFIA}
\title{Plot method for \code{tidyFIA} class}
\usage{
\method{plot}{tidyFIA}(x, ...)
}
\arguments{
\item{x}{object of class \code{tidyFIA} (output from \code{\link{tidy_fia}})}

\item{...}{Arguments to be passed \code{ggplot}}
}
\description{
Plot method for \code{tidyFIA} class
}
\author{
Henry Rodman
}
