% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_FIADB.R
\name{query_plot_table}
\alias{query_plot_table}
\title{query_plot_table}
\usage{
query_plot_table(aoi, con)
}
\arguments{
\item{aoi}{character state abbreviation code}

\item{con}{database connection}
}
\value{
sf object with FIA plot records
}
\description{
Perform spatial query on the plot table
}
\author{
Henry Rodman
}
