% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_FIADB.R
\name{query_table}
\alias{query_table}
\title{query_table}
\usage{
query_table(table_name, plt_cns, con)
}
\arguments{
\item{table_name}{'Oracle Table Name' as described in the 'Index of Tables' in FIADB
User Guide.}

\item{plt_cns}{character vector of FIA plot CNs from which to pull records}

\item{con}{connection to fiadb database}
}
\value{
dataframe with matching records from the target table
}
\description{
query_table
}
