% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{read_ref_table}
\alias{read_ref_table}
\title{Read FIA reference table}
\usage{
read_ref_table(table)
}
\arguments{
\item{table_name}{name of reference table e.g. "REF_SPECIES"}
}
\value{
dataframe of reference table
}
\description{
Read FIA reference table
}
