% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_tables}
\alias{ref_tables}
\title{ref_tables}
\format{
An object of class \code{list} of length 2.
}
\source{
data-raw/ref_tables.R
}
\usage{
ref_tables
}
\description{
list containing several FIADB reference tables: ref_species and 
ref_forest_type
}
\keyword{datasets}
