% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{stack_tables}
\alias{stack_tables}
\title{Stack tables}
\usage{
stack_tables(table_name, fia_db_files)
}
\arguments{
\item{table_name}{Oracle table name from FIADB}

\item{fia_db_files}{list of files downloaded by state}
}
\value{
dataframe containing data comibined from all states
}
\description{
Import all files called `table_name` from `fia_db_files` as
one dataframe
}
