% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{tidy_fia}
\alias{tidy_fia}
\title{Create tidy FIA tables}
\usage{
tidy_fia(states = NULL, aoi = NULL)
}
\arguments{
\item{states}{a character vector of state abbreviations, ignored if `aoi` is
supplied.}

\item{aoi}{sf object containing area of interest}
}
\value{
a list object containing tidy FIA data objects
}
\description{
This function queries the FIA database, by state abbreviation(s)
or area of interest,and returns a list of tidy data objects including the
TREE, PLOT, COND, and SURVEY tables.
}
\author{
Henry Rodman, Brian Clough
}
