% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{plot.tidyFIA}
\alias{plot.tidyFIA}
\title{Plot method for `tidyFIA` class}
\usage{
\method{plot}{tidyFIA}(x, ...)
}
\arguments{
\item{x}{object of class `tidyFIA` (output from `tidy_fia`)}

\item{...}{Arguments to be passed `ggplot`}
}
\description{
Plot method for `tidyFIA` class
}
\author{
Henry Rodman
}
