% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{tidy_fia}
\alias{tidy_fia}
\title{Create tidy FIA tables}
\usage{
tidy_fia(states = NULL, aoi = NULL, files = c("PLOT", "SUBPLOT",
  "COND", "TREE", "SURVEY"))
}
\arguments{
\item{states}{a character vector of state abbreviations, ignored if `aoi` is
supplied.}

\item{aoi}{sf object containing area of interest}

\item{files}{list of FIA tables to download. Tables must be identified by
their 'Oracle Table Name' as described in the 'Index of Tables' in FIADB
User Guide.}
}
\value{
a list object containing tidy data
}
\description{
This function queries the FIA database, by state abbreviation(s)
or area of interest,and returns a list of tidy data objects including the
TREE, PLOT, COND, and SURVEY tables.
}
\author{
Henry Rodman, Brian Clough
}
