% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_FIA.R
\name{tidy_fia}
\alias{tidy_fia}
\title{Create tidy FIA tables}
\usage{
tidy_fia(
  states = NULL,
  aoi = NULL,
  postgis = TRUE,
  table_names = c("plot", "subplot", "cond", "tree", "survey"),
  file_dir = tempdir()
)
}
\arguments{
\item{states}{a character vector of state abbreviations, ignored if
\code{aoi} is supplied.}

\item{aoi}{sf object containing area of interest}

\item{postgis}{logical if true, query PostGIS database instead of downloading
csvs from FIA Datamart.}

\item{table_names}{list of FIA tables to download. Tables must be identified
by their 'Oracle Table Name' as described in the 'Index of Tables' in FIADB
User Guide.}

\item{file_dir}{path to directory in which to download CSV files from the
Datamart. If the files are already present in this folder, the existing
files will be used.}
}
\value{
a list object containing tidy data
}
\description{
This function queries the FIA database, by state abbreviation(s)
or area of interest,and returns a list of tidy data objects including the
TREE, PLOT, COND, and SURVEY tables.
}
\author{
Henry Rodman, Brian Clough
}
